update stats params to match new api
