+++ libs/api-types/src/index.ts
@@ -9,6 +9,7 @@ export * from './report'
 export * from './resources'
 export * from './user-applications'
 export * from './user'
+export * from './stats'
 export * from './vessel'
 export * from './vesselGroups'
 export * from './workspaces'
